# public/openapi.yaml
openapi: 3.0.3
info:
  title: HookStep / Talent Platform API
  description: REST API for the hiring platform — jobs, applications, profile, and company dashboard.
  version: 1.0.0
servers:
  - url: /
    description: Current origin (relative)
...
paths:
...
  /api/profile:
    get:
      tags: [Profile]
      summary: Get my profile
      security: [{ session: [] }]
      responses:
        '200':
          description: Profile (no internal fields)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Profile]
      summary: Update my profile
      security: [{ session: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phoneNumber: { type: string }
                country: { type: string }
                state: { type: string }
                city: { type: string }
                linkedin: { type: string, format: uri }
                github: { type: string, format: uri }
                twitter: { type: string, format: uri }
                currentCTC: { type: string }
                expectedCTC: { type: string }
                noticePeriod: { type: string }
                profileImageUrl: { type: string, format: uri }
                resumeUrl: { type: string, format: uri }
                notificationNewApplications: { type: boolean }
                notificationInterviewUpdates: { type: boolean }
                notificationPlatformNews: { type: boolean }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Profile' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ----- Upload resume -----
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie (sign in via the app)
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
      headers: {}
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Profile:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        profileImageUrl: { type: string }
        phoneNumber: { type: string }
        country: { type: string }
        state: { type: string }
        city: { type: string }
        linkedin: { type: string }
        github: { type: string }
        twitter: { type: string }
        currentCTC: { type: string }
        expectedCTC: { type: string }
        noticePeriod: { type: string }
        resumeUrl: { type: string }
        notificationNewApplications: { type: boolean }
        notificationInterviewUpdates: { type: boolean }
        notificationPlatformNews: { type: boolean }